requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.24.0
pymongo==4.4.1
python-dotenv==1.0.0
requests==2.31.0
whisper==1.3.0
ffmpeg-python==0.2.0